tests/krate/publish: Extract `dependencies` module
